+++ demo/index.html
@@ -117,7 +117,7 @@
           s.parentNode.insertBefore(hm, s);
       })();
   </script>
-<script src="../dist/whatsElementWithUI.min.js"></script>
+<script src="../dist/whatsElement.js"></script>
 <script src="http://www.jq22.com/jquery/jquery-1.10.2.js"></script>
 <script src="highlight.pack.js"></script>
 <script>
