+++ extension/manifest.json
@@ -44,7 +44,7 @@
   "content_scripts": [
     {
       "matches": [ "<all_urls>" ],
-      "js": ["scripts/whatsElementWithUI.min.js","scripts/content_script.js"],
+      "js": ["scripts/whatsElement.min.js","scripts/content_script.js"],
       "all_frames": true
     }
   ]
