+++ index.js
@@ -1,5 +1,5 @@
 /**
  * Created by rowthan on 2017/11/19.
  */
-var whatsElement= require('./src/whatsElementWithUI')
+var whatsElement= require('./src/whatsElement')
 module.exports = whatsElement.default;
